+++ CHANGELOG.md
@@ -6,5 +6,6 @@ All notable changes will be tracked in this file. Most likely.
 - Added 3d6 Ability Score Generator
 - Added Point Buy Ability Score Generator
 - Added Ability to select generator via UI
+- Added Rogue Talents
 
 ### Changed